Added schemaOnly and AWS credentials support (#23)
* Added schemaOnly and AWS credentials support

The --schemaOnly option allows the user to backup and restore
just the schema information for the given table(s). The use
case for this is if you need to migrate your tables from one
AWS account to a different AWS account.

(Our specific use case: We have an AWS account for development
and yet a separate AWS account for production.)

To support the use case of multiple AWS acount, -p or --profile
was also added so that the user could explicitly specify which
profile to use from the AWS credentials file (~/.aws/credentials).
This is /sightly/ more secure than putting accessKey and
secretKey into the command line.

* Merged with upstream

* BUGFIX: Indent was wrong on merged line

Corrected error. Was accidentally running a version in my path instead
of the one in CWD... Sorry.