Confirm before deleting tables. (#51)
* Confirm before deleting tables.

The default options will delete and recreate tables during restores, which
can cause data loss that might be surprising. This adds a confirmation before
doing those deletes. It also adds a flag to disable confirmations for automated
cases (like tests or scripts).

* Update help output documentation in readme.

This is a copy/paste of the full output of python ./dynamodump.py -h.